requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
paho-mqtt>=2.0
flask>=3.0
flask-socketio>=5.3
